MedE commit

+++ src/main/java/com/MainProject/MedE/Store/StoreRegistrationModel.java
@@ -23,7 +23,7 @@ public class StoreRegistrationModel {
     private String licenseNumber;
 
     @Column(name = "phone_number")
-    private Integer phone_number;
+    private Long phone_number;
 
     @Lob
     @Column(name = "license_image")
@@ -85,11 +85,11 @@ public class StoreRegistrationModel {
         this.password = password;
     }
 
-    public Integer getPhone_number() {
+    public Long getPhone_number() {
         return phone_number;
     }
 
-    public void setPhone_number(Integer phone_number) {
+    public void setPhone_number(Long phone_number) {
         this.phone_number = phone_number;
     }
